Commands: Subjects

+++ LabsQueueBot.Web/Providers/INotificationProvider.cs
@@ -3,8 +3,7 @@
 public interface INotificationProvider
 {
     Task NotifyAll(CancellationToken cancellationToken);
-    Task NotifyGroup(short course, short group, CancellationToken cancellationToken);
-    Task NotifyGroupBySubject(short course, short group, string subjectName, List<long> queue, List<long> waiting, CancellationToken cancellationToken);
-    Task NotifyUserBySubject(long userId, string subjectName, CancellationToken cancellationToken);
-    Task NotifyAdminsWithDocument(int documentId, string message, CancellationToken cancellationToken);
+    Task NotifyGroup(byte course, byte group, CancellationToken cancellationToken);
+    Task NotifyGroupBySubject(byte course, byte group, string subjectName, List<long> queue, List<long> waiting, CancellationToken cancellationToken);
+    Task NotifyUserBySubject(long userId, string subjectName, CancellationToken cancellationToken); Task NotifyAdminsWithDocument(int documentId, string message, CancellationToken cancellationToken);
 }
